+++ lib/providers/preference_test_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:travel_muse_app/models/preference_test_model.dart';
-import 'package:travel_muse_app/services/preference_test_service.dart';
+import 'package:travel_muse_app/repositories/preference_test_repository.dart';
 import 'package:travel_muse_app/viewmodels/preference_test_view_model.dart';
 
 final preferenceTestProvider =
@@ -9,8 +9,8 @@ final preferenceTestProvider =
     );
 final preferenceTestStreamProvider =
     StreamProvider.family<PreferenceTest, String>((ref, testId) {
-      final service = PreferenceTestService();
-      return service.watchTest(testId);
+      final repository = PreferenceTestRepository();
+      return repository.watchTest(testId);
     });
 final preferenceTestViewModelProvider =
     NotifierProvider<PreferenceTestViewModel, AsyncValue<PreferenceTest?>>(
