+++ lib/repositories/preference_test_repository.dart
@@ -1,3 +1,4 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:travel_muse_app/models/preference_test_model.dart';
 import 'package:travel_muse_app/services/ai_service.dart';
 import 'package:travel_muse_app/services/preference_test_service.dart';
@@ -63,15 +64,42 @@ f: 새로운 것 탐험형 여행가
 
   Future<PreferenceTest> saveOrUpdateTest(PreferenceTest test) async {
     if (test.testId.isEmpty) {
-      final newId = await _service.saveTest(test);
+      final newId = await saveTest(test);
       return test.copyWith(testId: newId);
     } else {
-      await _service.saveOrUpdateTest(test);
+      await saveOrUpdateTest(test);
       return test;
     }
   }
 
   Future<PreferenceTest> loadTest(String testId) async {
-    return await _service.fetchTest(testId);
+    return await fetchTest(testId);
+  }
+
+  final _firestore = FirebaseFirestore.instance;
+  final _collection = 'preference_test';
+
+  /// Firestore에 테스트 저장
+  Future<String> saveTest(PreferenceTest test) async {
+    final docRef = _firestore.collection(_collection).doc();
+    final data = test.copyWith(testId: docRef.id).toMap();
+    await docRef.set(data);
+    return docRef.id;
+  }
+
+  /// Firestore에서 테스트 불러오기 (단건)
+  Future<PreferenceTest> fetchTest(String testId) async {
+    final doc = await _firestore.collection(_collection).doc(testId).get();
+    return PreferenceTest.fromDoc(doc.id, doc.data()!);
+  }
+
+  Stream<PreferenceTest> watchTest(String testId) {
+    return _firestore.collection(_collection).doc(testId).snapshots().map((
+      doc,
+    ) {
+      final data = doc.data();
+      if (data == null) throw Exception('문서가 존재하지 않음');
+      return PreferenceTest.fromDoc(doc.id, data);
+    });
   }
 }
